voice_defender_server/requirements1031.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
idna==3.4
mypy-extensions==1.0.0
opencv-python==4.8.1.78
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
py-cpuinfo==9.0.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
thop==0.1.1.post2209072238
tomli==2.0.1
typing_extensions==4.8.0
ultralytics==8.0.195
urllib3==2.0.6
uvicorn @ file:///C:/b/abs_bbigfycj8x/croot/uvicorn-split_1678090119838/work
